+++ test/js/local/cppBindingTest.js
@@ -59,21 +59,21 @@ describe('The native addon', () => {
 
     it('should be able to instantiate a TestClass using plain json', () => {
       const json = {
-        targetId: 'TestClass',
+        context: 'TestClass',
         method: '__create__',
         data: {} // No data => default ctor
       }
       const ret = JSON.parse(addon.call(JSON.stringify(json)))
       assert.property(ret, 'data')
       assert.isString(ret.data.r)
-      assert.property(ret, 'targetId')
+      assert.property(ret, 'context')
       assert.property(ret, 'method')
       instanceId = ret.data.r
     })
 
     it('should be able to call member function given valid instanceId', () => {
       const json = {
-        targetId: instanceId,
+        context: instanceId,
         method: 'hasCategory',
         data: { _1: 'test' }
       }
@@ -81,13 +81,13 @@ describe('The native addon', () => {
       assert.property(ret, 'data')
       assert.isBoolean(ret.data.r)
       assert.isFalse(ret.data.r)
-      assert.property(ret, 'targetId')
+      assert.property(ret, 'context')
       assert.property(ret, 'method')
     })
 
     it('should correctly handle call to non-existing function', () => {
       const json = {
-        targetId: instanceId,
+        context: instanceId,
         method: 'not_there',
         data: {}
       }
@@ -95,19 +95,19 @@ describe('The native addon', () => {
       assert.equal(ret.data.e, 'Could not find function: not_there')
     })
 
-    it('should correctly handle call to non-existing targetId', () => {
+    it('should correctly handle call to non-existing context', () => {
       const json = {
-        targetId: 'wrong',
+        context: 'wrong',
         method: 'not_there',
         data: {}
       }
       const ret = JSON.parse(addon.call(JSON.stringify(json)))
-      assert.equal(ret.data.e, 'Could not find targetId: wrong')
+      assert.equal(ret.data.e, 'Could not find context: wrong')
     })
 
     it('should properly trigger callbacks', () => {
       const json = {
-        targetId: instanceId,
+        context: instanceId,
         method: 'callMeBack',
         data: { _1: 'callback-1' }
       }
@@ -120,21 +120,21 @@ describe('The native addon', () => {
   describe('should properly handle named instances', () => {
     it('should be able to instantiate a TestClass using plain json', () => {
       const json = {
-        targetId: 'TestClass',
+        context: 'TestClass',
         method: '__createNamed__',
         data: { _1: 'test1' } // No data => default ctor
       }
       const ret = JSON.parse(addon.call(JSON.stringify(json)))
       assert.property(ret, 'data')
       assert.isString(ret.data.r)
-      assert.property(ret, 'targetId')
+      assert.property(ret, 'context')
       assert.property(ret, 'method')
       assert.strictEqual(ret.data.r, 'test1')
     })
 
     it('should be able to call member function given named instanceId', () => {
       const json = {
-        targetId: 'test1',
+        context: 'test1',
         method: 'hasCategory',
         data: { _1: 'test' }
       }
@@ -142,13 +142,13 @@ describe('The native addon', () => {
       assert.property(ret, 'data')
       assert.isBoolean(ret.data.r)
       assert.isFalse(ret.data.r)
-      assert.property(ret, 'targetId')
+      assert.property(ret, 'context')
       assert.property(ret, 'method')
     })
 
     it('should correctly handle call to non-existing function', () => {
       const json = {
-        targetId: 'test1',
+        context: 'test1',
         method: 'not_there',
         data: {}
       }
@@ -158,7 +158,7 @@ describe('The native addon', () => {
 
     it('should properly trigger callbacks', () => {
       const json = {
-        targetId: 'test1',
+        context: 'test1',
         method: 'callMeBack',
         data: { _1: 'callback-1' }
       }
@@ -169,7 +169,7 @@ describe('The native addon', () => {
 
     it('should properly delete the instance', () => {
       const json = {
-        targetId: 'TestClass',
+        context: 'TestClass',
         method: '__delete__',
         data: { _1: 'test1' }
       }
@@ -179,12 +179,12 @@ describe('The native addon', () => {
 
     it('should not be possible to call functions on a deleted instance', () => {
       const json = {
-        targetId: 'test1',
+        context: 'test1',
         method: 'hasCategory',
         data: { _1: 'test' }
       }
       const ret = JSON.parse(addon.call(JSON.stringify(json)))
-      assert.equal(ret.data.e, 'Could not find targetId: test1')
+      assert.equal(ret.data.e, 'Could not find context: test1')
     })
   })
 })